Move semaphore watch and decribe modes to separate classes
